Factor out install routine
